Move diff types to a separate module
